+++ cli/cmd/root.go
@@ -25,4 +25,5 @@ var rootCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(versionCmd)
 	rootCmd.AddCommand(pluginCmd)
+	rootCmd.AddCommand(mcpCmd)
 }